Created todo list and form with components that are instantized within App.js
